add nps summary response

+++ NPS.Core/Nps/ViewModels/NpsSummaryViewModel.cs
@@ -1,4 +1,4 @@
 namespace NPS.Core.Nps.ViewModels;
 
 // Relatório com as contagens de cada categoria e o cálculo do NPS
-public record NpsSummaryViewModel(int Promoters, int Neutrals, int Detractors, int NpsScore);
+public record NpsSummaryViewModel(decimal Promoters, decimal Neutrals, decimal Detractors, decimal NpsScore);
